change equation for count percentage of productivity,remove unnessesary code,add missed parts from develop
