+++ src/main/java/com/tstepnik/planner/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.tstepnik.planner.controller;
 
-import com.tstepnik.planner.domain.User;
+import com.tstepnik.planner.domain.user.User;
 import com.tstepnik.planner.security.auth.CustomUserDetails;
 import com.tstepnik.planner.service.UserService;
 import org.springframework.http.HttpStatus;
