--- src/main/java/com/tstepnik/planner/exceptions/ExpiredTaskException.java
@@ -1,7 +1,0 @@
-package com.tstepnik.planner.exceptions;
-
-public class ExpiredTaskException extends RuntimeException {
-    public ExpiredTaskException(String message) {
-        super(message);
-    }
-}
