--- src/test/java/com/tstepnik/planner/PlannerApplicationTests.java
@@ -1,8 +1,0 @@
-package com.tstepnik.planner;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PlannerApplicationTests {
-}
